refactor: EmojiBadge 폴더 위치 및 경로 수정 (#128)
