+++ src/components/EmojiBadge/EmojiBadge.jsx
