+++ src/pages/ListPage/components/RollingPaperCard/RollingPaperCard.jsx
@@ -1,5 +1,5 @@
 import ProfileStack from '@/components/ProfileStack/ProfileStack';
-import EmojiBadge from '@/EmojiBadge/EmojiBadge';
+import EmojiBadge from '@/components/EmojiBadge/EmojiBadge';
 import Divider from '@/components/Divider/Divider';
 
 // 서버 연결 후 지움
