+++ src/pages/PostedPage/PostedPage.jsx
@@ -5,7 +5,6 @@ import MessageCardList from './components/MessageCardList/MessageCardList';
 const PostedPage = () => {
   // const { id } = useParams();
   return (
-    //  <div className="grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-3 gap-6"></div>
     <div className="min-h-screen flex flex-col">
       {/* SubHeader 영역 */}
       <div className="sticky top-[64px] z-40 w-full ">
