+++ src/pages/PostedPage/components/SubHeader/SubHeader.jsx
@@ -1,5 +1,5 @@
 import Divider from '@/components/Divider/Divider';
-import EmojiBadge from '@/EmojiBadge/EmojiBadge';
+import EmojiBadge from '@/components/EmojiBadge/EmojiBadge';
 import ProfileStack from '@/components/ProfileStack/ProfileStack';
 import {
   Popover,
